categories added to context
