+++ src/util/ExampleData.tsx
@@ -66,6 +66,7 @@ export interface Conversation {
 export interface Category {
   id: number;
   name: string;
+  color?: string;
   colorCode: TagColor;
 }
